pyproject.toml
[tool.poetry]
name = "qrcode-scanner"
version = "1.0.0"
description = "Software that make you can decode QR code from image."
authors = ["oxygen-TW <oxygen5307@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
pyzbar = "opencv-python"
opencv-python = "^4.6.0"
clipboard = "^0.0.4"
Pillow = "^9.2.0"

[tool.poetry.dev-dependencies]
pyinstaller = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
